Issue #100 Refactor modules from one-class-per-file to one-module-per-file
