--- polls/controllers/__init__.py
@@ -1,1 +1,0 @@
-from .vote_controller import VoteController
