+++ polls/models.py
@@ -1,5 +1,13 @@
 from django.db import models
-from polls.models.question import Question
+
+
+class Question(models.Model):
+    question_text = models.CharField(max_length=200)
+    pub_date = models.DateTimeField('date published')
+
+    def __str__(self):
+        return self.question_text
+
 
 class Choice(models.Model):
