--- polls/models/__init__.py
@@ -1,2 +1,0 @@
-from .choice import Choice
-from .question import Question
