--- polls/repositories/__init__.py
@@ -1,1 +1,0 @@
-from .choice_repository import ChoiceRepository
