--- polls/views/__init__.py
@@ -1,6 +1,0 @@
-from .index_view import IndexView
-from .detail_view import DetailView
-from .results_view import ResultsView
-from .vote import vote
-from .question_view_set import QuestionViewSet
-from .choice_view_set import ChoiceViewSet
