--- polls/views/question_view_set.py
@@ -1,7 +1,0 @@
-from rest_framework import viewsets
-from polls import models
-from polls import serializers
-
-class QuestionViewSet(viewsets.ModelViewSet):
-    queryset = models.Question.objects.all()
-    serializer_class = serializers.QuestionSerializer
