--- polls/views/results_view.py
@@ -1,7 +1,0 @@
-from django.views import generic
-from polls import models
-
-# pylint: disable=too-many-ancestors
-class ResultsView(generic.DetailView):
-    model = models.Question
-    template_name = 'polls/results.html'
