+++ polls/viewsets.py
@@ -2,9 +2,12 @@ from rest_framework import viewsets
 from polls import models
 from polls import serializers
 
+class QuestionViewSet(viewsets.ModelViewSet):
+    queryset = models.Question.objects.all()
+    serializer_class = serializers.QuestionSerializer
+
 
 class ChoiceViewSet(viewsets.ModelViewSet):
-
     serializer_class = serializers.ChoiceSerializer
 
     def get_queryset(self):
